Overrides for 1.5 mostly in place, need to add Gary's updates [#8] Inline documentation updates [#20]

+++ trunk/wright/parameters/joomla_1.5/help.php
@@ -14,7 +14,7 @@ class JElementHelp extends JElement
 		$html = '<a class="modal" href="'.JURI::root().'templates/'.$template.'/wright/help" rel="{\'handler\': \'iframe\', \'size\': {x: 800, y:600}}">'.JText::_('Documentation').'</a>';
 
 		// Refresh CSS cache since we are editing params
-		JFile::delete(JPATH_ROOT.DS.'templates'.DS.$template.DS.'css'.DS.$template.'.css');
+		if (is_file(JPATH_ROOT.DS.'templates'.DS.$template.DS.'css'.DS.$template.'.css')) JFile::delete(JPATH_ROOT.DS.'templates'.DS.$template.DS.'css'.DS.$template.'.css');
 
 		return $html;
 	}
