+++ trunk/wright/parameters/joomla_1.6/help.php
@@ -10,14 +10,12 @@ class JFormFieldHelp extends JFormField
 
 	protected function getInput($name, $value, &$node, $control_name)
 	{
-		
-
 		JHTML::_('behavior.modal');
 		$template = $_GET['cid'][0];
 		$html = '<a class="modal" href="'.JURI::root().'templates/'.$template.'/wright/help" rel="{\'handler\': \'iframe\', \'size\': {x: 800, y:600}}">'.JText::_('Documentation').'</a>';
 
 		// Refresh CSS cache since we are editing params
-		JFile::delete(JPATH_ROOT.DS.'templates'.DS.$template.DS.'css'.DS.$template.'.css');
+		if (is_file(JPATH_ROOT.DS.'templates'.DS.$template.DS.'css'.DS.$template.'.css')) JFile::delete(JPATH_ROOT.DS.'templates'.DS.$template.DS.'css'.DS.$template.'.css');
 
 		return $html;
 	}
